+++ CHANGELOG.md
@@ -127,6 +127,18 @@
 
 
 
+ * Appsec
+  * Added method `GetConfigurationVersion` to get specific config version, to fix a bug where TF provider doesn't create new config version but tries to update in place.([I#653](https://github.com/akamai/terraform-provider-akamai/issues/653)).
+
+
+
+
+
+
+
+
+
+